Delay-load icu so that we don't fail to start up on Windows <1903
Originally, I intended to switch to icuuc/icuin.dll to make it work on
Windows 1703 as well... but it turns out that the 22621 or 26100 SDKs
redirect those back to icu.dll. That's a poor decision on their part,
as it means that developers can't use the old compatibility names.

VS supports Windows <=1903, so we need to be able to load. Fortunately,
the aren't using any of the features that need ICU (search and URL
detection). Therefore, we don't need to guard our ICU calls... yet.

ConPTY doesn't use them at all.

When we add support for URL detection and search to HwndTerminal, we'll
need to revisit.