Restore the alias too?
this makes resetting the background color the opposite of setting it:
setting it changes the alias to DEFAULT_BACKGROUND, and resetting it
changes it back to INDEX=0 (conhost).

As a side effect, the value of DEFAULT_BACKGROUND is also destroyed.

One could test this in xterm by using DECAC to set the background to
their internal background color index and observing whether it, too, was
preserved.