wpf: strong-name sign Microsoft.Terminal.Wpf (#18836)
This requires us to delay-sign the assembly with a public key (the snk
file), and then later submit it for strong naming. This is separate from
code signing, and has to take place before it.

The snk file does not contain any private key material.

This cannot merge until we are approved to use this new signing "key
code".

(cherry picked from commit 0568173abab4150d167dcced3994c99356dfc5cd)
Service-Card-Id: PVTI_lADOAF3p4s4AmhmQzgZuEzY
Service-Version: 1.22